requirements.txt
monai==1.5.0
torch==2.6.0
numpy==2.3.2
pandas==2.3.2
psycopg2-binary==2.9.10
simpleitk==2.5.2
